Day 9: Bugs are fine
My solution for part 1 actually has a bug in it. It also checks the
diagonals for the lowest point. Therefore, if there are lowest points on
the diagonal this solution won't find them. However, due to part 2
requiring the basins, the lowest points are spread out and thus didn't
find this bug.