Day 19: Triangulation >:(
Yesterdays puzzle is nothing compared to this one. It took me way to
long to solve this. Probably the most time is spend in rotating the
scanners. I tried to minimize the manual labour with it but then I
figured out that you cannot generate a new list of lambda's with two
other lists of lambda's. Therefore, I still had to manually write down
all the rotations. Otherwise, the puzzle isn't that bad.