Day 23: Always use graphs!
I tried to do part 1 with hand. However, I probable miscounted as that
failed. I didn't know how to model this problem so again I had to search
for hits (Shame!). Someone suggested to model it as a graph and find the
shortest path. So I solved the problem using bruteforce with
memorization and it is reasonably quick. The code is quite ugle however.