Day 24: My first keygen
This puzzle involves a bit of reverse engineering which is a bit of fun.
I also implemented the logic of the ALU but that was not necessary.
However, it was useful for verifying if my code works. Furthermore again
you needed to do a search for the largest and smallest value. I'm a bit
tired of writing a search function.